+++ compiler/lib/code.ml
@@ -122,6 +122,8 @@ module Var : sig
 
     val set : 'a t -> key -> 'a -> unit
 
+    val length : 'a t -> int
+
     val make : size -> 'a -> 'a t
 
     val make_map : size -> ('a, 'b) DataMap.t t
@@ -271,6 +273,8 @@ end = struct
     let get t x = t.(x)
 
     let set t x v = t.(x) <- v
+
+    let length t = Array.length t
 
     let make () v = Array.make (count ()) v