+++ compiler/lib/code.mli
@@ -115,6 +115,8 @@ module Var : sig
 
     val set : 'a t -> key -> 'a -> unit
 
+    val length : 'a t -> int
+
     val make : size -> 'a -> 'a t
 
     val make_map : size -> ('a, 'b) DataMap.t t