+++ compiler/lib/effects.mli
@@ -20,8 +20,21 @@ type cps_calls = Code.Var.Set.t
 
 val remove_empty_blocks : live_vars:Deadcode.variable_uses -> Code.program -> Code.program
 
+type single_version_closures = Code.Var.Set.t
+
 val f :
      flow_info:Global_flow.info
   -> live_vars:Deadcode.variable_uses
   -> Code.program
-  -> Code.program * cps_calls
+  -> Code.program * cps_calls * single_version_closures
+(** Perform a partial CPS transform to translate the result effect handlers.
+
+    In addition, if the [doubletranslate] feature is enabled, functions are
+    created in two versions (direct-style and CPS) and the generated program
+    switches to CPS when entering the first effect handler, and back to direct
+    style when exiting it. In addition to this dynamic behavior, the transform
+    performs a static analysis to detect which functions do not need to be
+    CPS-transformed. As a consequence, some functions become pairs of functions
+    while others remain in a single version. This functions returns the
+    locations of CPS calls, and the set of closures that exist in a single
+    version (which can be in direct style or CPS). *)