+++ compiler/lib/generate.ml
@@ -56,6 +56,7 @@ type application_description =
   { arity : int
   ; exact : bool
   ; cps : bool
+  ; single_version : bool
   }
 
 module Share = struct
@@ -134,6 +135,7 @@ module Share = struct
 
   let get
       ~cps_calls
+      ~single_version_closures
       ?alias_strings
       ?(alias_prims = false)
       ?(alias_apply = true)
@@ -151,8 +153,15 @@ module Share = struct
               | Let (_, Constant c) -> get_constant c share
               | Let (x, Apply { args; exact; _ }) ->
                   let cps = Var.Set.mem x cps_calls in
+                  let single_version =
+                    (not (Config.Flag.double_translation ()))
+                    || Var.Set.mem x single_version_closures
+                  in
                   if (not exact) || cps
-                  then add_apply { arity = List.length args; exact; cps } share
+                  then
+                    add_apply
+                      { arity = List.length args; exact; cps; single_version }
+                      share
                   else share
               | Let (_, Special (Alias_prim name)) ->
                   let name = Primitive.resolve name in
@@ -244,15 +253,22 @@ module Share = struct
       try J.EVar (AppMap.find desc t.vars.applies)
       with Not_found ->
         let x =
-          let { arity; exact; cps } = desc in
+          let { arity; exact; cps; single_version } = desc in
           Var.fresh_n
             (Printf.sprintf
                "caml_%scall%d"
-               (match exact, cps with
-               | true, false -> assert false
-               | true, true -> "cps_exact_"
-               | false, false -> ""
-               | false, true -> "cps_")
+               (match exact, cps, single_version with
+               | true, true, false -> "cps_exact_double_"
+               | true, true, true -> "cps_exact_"
+               | false, false, false -> "double"
+               | false, false, true -> ""
+               | false, true, false -> "cps_"
+               | false, true, true ->
+                   assert (not (Config.Flag.double_translation ()));
+                   "cps_"
+               | true, false, _ ->
+                   (* Should not happen: no intermediary function needed *)
+                   assert false)
                arity)
         in
         let v = J.V x in
@@ -273,6 +289,7 @@ module Ctx = struct
     ; deadcode_sentinal : Var.t
     ; mutated_vars : Code.Var.Set.t Code.Addr.Map.t
     ; freevars : Code.Var.Set.t Code.Addr.Map.t
+    ; single_version_closures : Effects.single_version_closures
     }
 
   let initial
@@ -285,6 +302,7 @@ module Ctx = struct
       blocks
       live
       cps_calls
+      single_version_closures
       share
       debug =
     { blocks
@@ -298,6 +316,7 @@ module Ctx = struct
     ; deadcode_sentinal
     ; mutated_vars
     ; freevars
+    ; single_version_closures
     }
 end
 
@@ -773,49 +792,74 @@ let parallel_renaming back_edge params args continuation queue =
 
 (****)
 
-let apply_fun_raw ctx f params exact cps =
-  let n = List.length params in
-  let apply_directly =
-    (* Make sure we are performing a regular call, not a (slower)
-       method call *)
-    match f with
-    | J.EAccess _ | J.EDot _ ->
-        J.call (J.dot f (Utf8_string.of_string_exn "call")) (s_var "null" :: params) J.N
-    | _ -> J.call f params J.N
-  in
-  let apply =
-    (* We skip the arity check when we know that we have the right
-       number of parameters, since this test is expensive. *)
-    if exact
-    then apply_directly
-    else
-      let l = Utf8_string.of_string_exn "l" in
+let apply_fun_raw =
+  let cps_field = Utf8_string.of_string_exn "cps" in
+  fun ctx f params exact cps single_version ->
+    let n = List.length params in
+    let apply_directly f params =
+      (* Make sure we are performing a regular call, not a (slower)
+         method call *)
+      match f with
+      | J.EAccess _ | J.EDot _ ->
+          J.call (J.dot f (Utf8_string.of_string_exn "call")) (s_var "null" :: params) J.N
+      | _ -> J.call f params J.N
+    in
+    let apply cps single =
+      (* Adapt if [f] is a (direct-style, CPS) closure pair *)
+      let real_closure =
+        if (not (Config.Flag.effects ())) || (not cps) || single
+        then f
+        else
+          (* Effects enabled, CPS version, not single-version *)
+          J.EDot (f, J.ANormal, cps_field)
+      in
+      (* We skip the arity check when we know that we have the right
+         number of parameters, since this test is expensive. *)
+      if exact
+      then apply_directly real_closure params
+      else
+        let l = Utf8_string.of_string_exn "l" in
+        J.ECond
+          ( J.EBin
+              ( J.EqEq
+              , J.ECond
+                  ( J.EBin (J.Ge, J.dot real_closure l, int 0)
+                  , J.dot real_closure l
+                  , J.EBin
+                      ( J.Eq
+                      , J.dot real_closure l
+                      , J.dot real_closure (Utf8_string.of_string_exn "length") ) )
+              , int n )
+          , apply_directly real_closure params
+          , J.call
+              (* Note: when double translation is enabled, [caml_call_gen*] functions takes a two-version function *)
+              (runtime_fun
+                 ctx
+                 (if cps && Config.Flag.double_translation ()
+                  then "caml_call_gen_cps"
+                  else "caml_call_gen"))
+              [ f; J.array params ]
+              J.N )
+    in
+    if cps
+    then (
+      assert (Config.Flag.effects ());
+      (* When supporting effect, we systematically perform tailcall
+         optimization. To implement it, we check the stack depth and
+         bounce to a trampoline if needed, to avoid a stack overflow.
+         The trampoline then performs the call in an shorter stack. *)
+      let f =
+        if single_version && Config.Flag.double_translation ()
+        then J.(EObj [ Property (PNS (Utf8_string.of_string_exn "cps"), f) ])
+        else f
+      in
       J.ECond
-        ( J.EBin
-            ( J.EqEq
-            , J.ECond
-                ( J.EBin (J.Ge, J.dot f l, int 0)
-                , J.dot f l
-                , J.EBin (J.Eq, J.dot f l, J.dot f (Utf8_string.of_string_exn "length"))
-                )
-            , int n )
-        , apply_directly
-        , J.call (runtime_fun ctx "caml_call_gen") [ f; J.array params ] J.N )
-  in
-  if cps
-  then (
-    assert (Config.Flag.effects ());
-    (* When supporting effect, we systematically perform tailcall
-       optimization. To implement it, we check the stack depth and
-       bounce to a trampoline if needed, to avoid a stack overflow.
-       The trampoline then performs the call in an shorter stack. *)
-    J.ECond
-      ( J.call (runtime_fun ctx "caml_stack_check_depth") [] J.N
-      , apply
-      , J.call (runtime_fun ctx "caml_trampoline_return") [ f; J.array params ] J.N ))
-  else apply
+        ( J.call (runtime_fun ctx "caml_stack_check_depth") [] J.N
+        , apply cps single_version
+        , J.call (runtime_fun ctx "caml_trampoline_return") [ f; J.array params ] J.N ))
+    else apply cps single_version
 
-let generate_apply_fun ctx { arity; exact; cps } =
+let generate_apply_fun ctx { arity; exact; cps; single_version } =
   let f' = Var.fresh_n "f" in
   let f = J.V f' in
   let params =
@@ -830,10 +874,13 @@ let generate_apply_fun ctx { arity; exact; cps } =
     ( None
     , J.fun_
         (f :: params)
-        [ J.Return_statement (Some (apply_fun_raw ctx f' params' exact cps)), J.N ]
+        [ ( J.Return_statement
+              (Some (apply_fun_raw ctx f' params' exact cps single_version))
+          , J.N )
+        ]
         J.N )
 
-let apply_fun ctx f params exact cps loc =
+let apply_fun ctx f params exact cps single_version loc =
   (* We always go through an intermediate function when doing CPS
      calls. This function first checks the stack depth to prevent
      a stack overflow. This makes the code smaller than inlining
@@ -841,12 +888,12 @@ let apply_fun ctx f params exact cps loc =
      since the function should get inlined by the JavaScript
      engines. *)
   if Config.Flag.inline_callgen () || (exact && not cps)
-  then apply_fun_raw ctx f params exact cps
+  then apply_fun_raw ctx f params exact cps single_version
   else
     let y =
       Share.get_apply
         (generate_apply_fun ctx)
-        { arity = List.length params; exact; cps }
+        { arity = List.length params; exact; cps; single_version }
         ctx.Ctx.share
     in
     J.call y (f :: params) loc
@@ -1030,6 +1077,10 @@ let rec translate_expr ctx queue loc x e level : _ * J.statement_list =
   match e with
   | Apply { f; args; exact } ->
       let cps = Var.Set.mem x ctx.Ctx.cps_calls in
+      let single_version =
+        (not (Config.Flag.double_translation ()))
+        || Var.Set.mem f ctx.Ctx.single_version_closures
+      in
       let args, prop, queue =
         List.fold_right
           ~f:(fun x (args, prop, queue) ->
@@ -1040,7 +1091,7 @@ let rec translate_expr ctx queue loc x e level : _ * J.statement_list =
       in
       let (prop', f), queue = access_queue queue f in
       let prop = or_p prop prop' in
-      let e = apply_fun ctx f args exact cps loc in
+      let e = apply_fun ctx f args exact cps single_version loc in
       (e, prop, queue), []
   | Block (tag, a, array_or_not, _mut) ->
       let contents, prop, queue =
@@ -1446,7 +1497,8 @@ and translate_instrs_rev (ctx : Ctx.t) expr_queue instrs acc_rev muts_map : _ * 
             List.fold_left
               pcs
               ~init:(ctx.blocks, Addr.Set.empty)
-              ~f:(fun (blocks, visited) pc -> Subst.cont' subst pc blocks visited)
+              ~f:(fun (blocks, visited) pc ->
+                Subst.Excluding_Binders.cont' subst pc blocks visited)
           in
           { ctx with blocks = p }
       in
@@ -1950,12 +2002,15 @@ let f
     ~exported_runtime
     ~live_vars
     ~cps_calls
+    ~single_version_closures
     ~should_export
     ~warn_on_unhandled_effect
     ~deadcode_sentinal
     debug =
   let t' = Timer.make () in
-  let share = Share.get ~cps_calls ~alias_prims:exported_runtime p in
+  let share =
+    Share.get ~cps_calls ~single_version_closures ~alias_prims:exported_runtime p
+  in
   let exported_runtime =
     if exported_runtime then Some (Code.Var.fresh_n "runtime", ref false) else None
   in
@@ -1972,6 +2027,7 @@ let f
       p.blocks
       live_vars
       cps_calls
+      single_version_closures
       share
       debug
   in